Large_factorial
We can find a large number of factorial with the help of a vector or array, cus it is not possible to store such a large number in C/C++.For example, the Factorial of 100 has 158 digits. It is not possible to store these many digits even if we use long long int.